chardev: Let qemu_chr_be_* use size_t type
Since IOCanReadHandler now returns an unsigned value, the assertion
does not make sense anymore.  We choose however to keep a "fail-safe"
assertion to catch undesired overflows.

Suggested-by: Paolo Bonzini <pbonzini@redhat.com>

+++ chardev/char-socket.c
@@ -488,7 +488,7 @@ static gboolean tcp_chr_read(QIOChannel *chan, GIOCondition cond, void *opaque)
             tcp_chr_process_IAC_bytes(chr, s, buf, &size);
         }
         if (size > 0) {
-            qemu_chr_be_write(chr, buf, size);
+            qemu_chr_be_write(chr, buf, (size_t)size);
         }
     }
